add Pundit authorization to BreweryController

+++ app/controllers/breweries_controller.rb
@@ -28,6 +28,8 @@ class BreweriesController < ApplicationController
   def update
     # patch /breweries/:id will go to breweries#update and edit a brewery
     brewery = Brewery.find(params[:id])
+    # Pundit authorization checking user.admin == true
+    authorize brewery 
     if brewery.update(brewery_params)
       render json: brewery 
     else
@@ -38,6 +40,8 @@ class BreweriesController < ApplicationController
   def destroy
     # post with method of delete /breweries/:id will go to breweries#destroy and delete a brewery
     brewery = Brewery.find(params[:id])
+    # Pundit authorization checking user.admin == true
+    authorize brewery
     brewery.destroy
   end
